refactor: split the conformance driver into 4 files
Per review: the ~1000-line notte/conformance.py is now conformance.py (entry:
stepper loop, browser backends, baseline gate, CLI), chatgpt.py and claude.py
(host adapters), and utils.py (result models, HostConfig, shared widget and
overlay helpers). Sibling imports resolve because uv puts the entry script's
dir on sys.path; the inline deps stay in conformance.py. README documents the
four files.
