+++ examples/conformance/notte/conformance.py
@@ -4,18 +4,21 @@
 #     "playwright>=1.48",
 #     "pydantic>=2",
 #     "notte-sdk",
+#     "python-dotenv>=1",
 # ]
 # ///
 """
 Drive the Skybridge hooks conformance app end-to-end on a real host with plain
-Playwright (sync API).
+Playwright (sync API). This is the entry point: it owns the stepper loop, the
+browser backends, the baseline gate, and the CLI. Host-specific behavior lives
+in chatgpt.py / claude.py; shared helpers and models live in utils.py.
 
 Sends a single "run @{app_name}" (ChatGPT) / "run {app_name}" (Claude)
 prompt, then drives the in-widget stepper like a real tester would: clicks
 Run on each action test, verifies the side effects it can observe from
-outside the widget (modal, new tab, follow-up message) to answer the app's
-Yes/No confirmations honestly, skips the ones it cannot observe, extracts the
-results table, and presses Close last.
+outside the widget (modal, follow-up message, native dialogs) to answer the
+app's Yes/No confirmations honestly, skips the ones it cannot observe,
+extracts the results table, and presses Close last.
 
 Two browser backends are supported:
 
@@ -41,534 +44,36 @@ import os
 import sys
 import time
 from contextlib import contextmanager
-from dataclasses import dataclass
 from pathlib import Path
-from typing import Any, Callable, Iterator
+from typing import Any, Iterator
 
-from pydantic import BaseModel
+from dotenv import load_dotenv
 from playwright.sync_api import BrowserContext, Page, sync_playwright
 
-
-class ResultRow(BaseModel):
-    hook: str
-    result: str
-    detail: str
-
-
-class StepperState(BaseModel):
-    run_complete: bool
-    current_hook: str | None
-    action_button: str | None
-    confirm_question: str | None
-    modal_open: bool
-    rows: list[ResultRow]
-
-
-class StallError(Exception):
-    """A test stopped advancing (e.g. ChatGPT remounts the iframe on fullscreen,
-    dropping the displayMode test's armed state), so the run can't make progress
-    and should be retried in a fresh conversation."""
-
-
-# Verdicts that mean a row hasn't resolved yet.
-_PENDING = ("", "testing", "testing…")
-
-
-def rows_complete(rows: list[ResultRow]) -> bool:
-    """Every test broadcast a real verdict (nothing left pending)."""
-    return bool(rows) and all(r.result.strip() not in _PENDING for r in rows)
-
-
-# Timeout (ms) for the first action after navigation (SPA hydration).
-PAGE_LOAD_TIMEOUT_MS = 30_000
-
-# Stepper safety rails: the app has 15 tests, and useRegisterViewTool waits a
-# full 60s for the host to invoke the view tool (it never does, so it times out
-# to unsupported), which alone is ~15 idle turns; a stuck widget still must not
-# spin forever.
-MAX_LOOP_TURNS = 60
-STEPPER_DEADLINE_S = 10 * 60
-
-FOLLOW_UP_MARKER = "Skybridge conformance test"
-
-
-# ── Host adapters ────────────────────────────────────────────────────
-
-
-@dataclass
-class HostConfig:
-    name: str
-    url: str
-    widget_iframe_selector: str
-    send_prompt: Callable[[Page, str], None]
-    hide_sidebar: Callable[[Page], None]
-    dismiss_modal: Callable[[Page], None]
-    check_follow_up: Callable[[Page], bool]
-    # Some hosts gate a hook's effect behind a native permission dialog that
-    # must be accepted for the effect to happen (Claude prompts "Open link"
-    # for openExternal and "Download" for download). Called right after Run
-    # with the current hook; returns True/False when it owns the verification
-    # (accepted/failed), or None to let the generic side-effect checks run.
-    accept_native_dialog: Callable[[Page, str], bool | None] | None = None
-
-
-def send_prompt_chatgpt(page: Page, app_name: str) -> None:
-    """Send the single "run @{app_name}" prompt.
-
-    Typing @name pops ChatGPT's app-mention picker; the first Enter selects
-    the app from the picker, the second sends the message.
-    """
-    page.fill("#prompt-textarea", f"run @{app_name}", timeout=PAGE_LOAD_TIMEOUT_MS)
-    time.sleep(3)  # mention picker
-    page.keyboard.press("Enter")
-    time.sleep(1)
-    page.keyboard.press("Enter")
-
-
-def hide_sidebar_chatgpt(page: Page) -> None:
-    """Collapse the ChatGPT left sidebar so the widget fills more of the screenshot."""
-    page.evaluate(
-        """() => {
-            const btn = document.querySelector('button.cursor-w-resize');
-            if (btn) btn.click();
-        }"""
-    )
-    time.sleep(1)
-
-
-def dismiss_modal_chatgpt(page: Page) -> None:
-    """Dismiss any blocking ChatGPT modal (e.g. rate-limit dialog)."""
-    page.evaluate(
-        """() => {
-            const btn = [...document.querySelectorAll('button')].find(b => b.textContent.trim() === 'Got it');
-            if (btn) btn.click();
-        }"""
-    )
-    time.sleep(1)
-
-
-def check_follow_up_chatgpt(page: Page, timeout_seconds: int = 120, poll_interval: int = 8) -> bool:
-    """Detect the follow-up message in the conversation itself.
-
-    The conversation DOM is not a reliable witness: the widget stays
-    fullscreen after the displayMode test (ChatGPT ignores widget-initiated
-    restore), hiding the transcript. Query ChatGPT's same-origin backend API
-    for the conversation content instead. Unlike the Notte version of this
-    driver (which could only read a markdown-string return value and had to
-    kick the fetch off and poll a window variable), plain Playwright can just
-    await the promise directly inside evaluate().
-
-    The polling window is long on purpose: the snapshot endpoint only
-    reflects a turn once it completes, which was observed to lag the
-    dispatch by 45s+ (network logs of run 3e389d74 show the follow-up POST at
-    t+0 and the conversation record still unchanged at t+38s).
-    """
-    js = f"""
-    async () => {{
-        const convId = (location.pathname.match(/\\/c\\/([a-z0-9-]+)/i) || [])[1];
-        if (!convId) return 'no-conversation-id';
-        const sessResp = await fetch('/api/auth/session');
-        const token = (await sessResp.json()).accessToken;
-        const accountId = (document.cookie.match(/_account=([^;]+)/) || [])[1] || '';
-        const resp = await fetch('/backend-api/conversation/' + convId, {{
-            headers: {{
-                'Authorization': 'Bearer ' + token,
-                'ChatGPT-Account-ID': accountId,
-            }},
-        }});
-        if (!resp.ok) return 'http-' + resp.status;
-        const text = JSON.stringify(await resp.json());
-        return text.includes({json.dumps(FOLLOW_UP_MARKER)}) ? 'found' : 'not-found';
-    }}
-    """
-    elapsed = 0
-    while elapsed < timeout_seconds:
-        try:
-            status = page.evaluate(js)
-        except Exception as exc:
-            status = f"error: {exc}"
-        print(f"[conformance] follow-up check: {status}", flush=True)
-        if status == "found":
-            return True
-        time.sleep(poll_interval)
-        elapsed += poll_interval
-    return False
-
-
-def accept_native_dialog_chatgpt(page: Page, hook: str) -> bool | None:
-    """Accept ChatGPT's native permission dialog so the effect happens.
-
-    openExternal pops an "External site" dialog (Copy link / Open link);
-    clicking "Open link" confirms the host handled the hook and dismisses the
-    backdrop. (click_top_page_button is defined below; resolved at call time.)
-    """
-    if "openexternal" in hook.lower():
-        return click_top_page_button(page, "Open link")
-    return None
-
-
-CHATGPT = HostConfig(
-    name="chatgpt",
-    url="https://chatgpt.com/",
-    widget_iframe_selector='iframe[src*="oaiusercontent"]',
-    send_prompt=send_prompt_chatgpt,
-    hide_sidebar=hide_sidebar_chatgpt,
-    dismiss_modal=dismiss_modal_chatgpt,
-    check_follow_up=check_follow_up_chatgpt,
-    accept_native_dialog=accept_native_dialog_chatgpt,
-)
-
-
-def send_prompt_claude(page: Page, app_name: str) -> None:
-    """Send the single "run {app_name}" prompt (verified live).
-
-    Claude's composer is a ProseMirror `contenteditable` div, not an
-    `<input>`/`<textarea>`. No mention picker: a single Enter sends, and
-    Claude resolves the app by name from the connected MCP servers. Both the
-    typing and the Enter get silently swallowed sometimes (cookie banner /
-    onboarding popovers stealing focus), so verify each and retry.
-    """
-    dismiss_modal_claude(page)  # the cookie banner overlay steals the first click
-    prompt = f"run {app_name}"
-    composer = page.locator('div[contenteditable="true"]').first
-    composer.click(timeout=PAGE_LOAD_TIMEOUT_MS)
-    for _ in range(3):
-        composer.fill(prompt)
-        time.sleep(1)
-        if prompt in (composer.inner_text() or ""):
-            break
-    for _ in range(5):
-        page.keyboard.press("Enter")
-        time.sleep(3)
-        # a sent message navigates to /chat/<id> and empties the composer
-        if "/chat/" in page.url:
-            return
-        try:
-            if prompt not in (composer.inner_text() or ""):
-                return
-        except Exception:
-            return  # composer re-rendered away: the message left
-    raise TimeoutError("the claude composer never sent the prompt")
-
-
-def hide_sidebar_claude(page: Page) -> None:
-    """No-op: Claude's sidebar doesn't overlap the widget, nothing to collapse."""
-    return
-
-
-def dismiss_modal_claude(page: Page) -> None:
-    """Accept the cookie banner: its overlay swallows clicks near the composer."""
-    page.evaluate(
-        """() => {
-            const btn = [...document.querySelectorAll('button')].find(b => b.textContent.trim() === 'Accept All Cookies');
-            if (btn) btn.click();
-        }"""
-    )
-    time.sleep(1)
-
-
-def click_top_page_button(page: Page, label: str, timeout_seconds: int = 20) -> bool:
-    """Click a control by exact text in a top-page host dialog (not the widget).
-
-    Hosts render native permission dialogs (Open link / Download) as top-page
-    portals that appear a beat after the triggering hook runs and block all
-    other clicks until answered; poll for the control, then click it with the
-    same CDP-aware timeout as real_click (a 2s click times out over Notte's
-    remote browser, which silently left the download unaccepted).
-
-    The control's role varies by host: Claude's "Open link" is a <button>, but
-    ChatGPT's "External site" dialog renders "Open link" as an <a> — so match
-    button, link, or bare text rather than assuming a role.
-    """
-    deadline = time.time() + timeout_seconds
-    while time.time() < deadline:
-        for locator in (
-            page.get_by_role("button", name=label, exact=True),
-            page.get_by_role("link", name=label, exact=True),
-            page.get_by_text(label, exact=True),
-        ):
-            try:
-                if locator.count():
-                    locator.first.click(timeout=CLICK_TIMEOUT_MS)
-                    return True
-            except Exception:
-                continue
-        time.sleep(1)
-    return False
-
-
-def accept_native_dialog_claude(page: Page, hook: str) -> bool | None:
-    """Accept Claude's native permission dialogs so the effect actually happens.
-
-    openExternal and download are gated behind a host dialog ("Open link" /
-    "Download"); without a click the tab never opens and download() times out,
-    and the dialog's backdrop blocks every later click. requestModal renders
-    the view in modal mode inside the widget (no host dialog) and the app's
-    own display-store check reports it, matching the manual baseline.
-    """
-    hook_lc = hook.lower()
-    if "openexternal" in hook_lc:
-        return click_top_page_button(page, "Open link")
-    if "download" in hook_lc:
-        return click_top_page_button(page, "Download")
-    if "requestmodal" in hook_lc:
-        return True
-    return None
-
-
-def check_follow_up_claude(page: Page, timeout_seconds: int = 120, poll_interval: int = 8) -> bool:
-    """TODO(claude): verify live.
-
-    There's no known same-origin backend API to query on claude.ai (unlike
-    ChatGPT's /backend-api/conversation/<id>), so this falls back to scraping
-    the visible conversation transcript for the marker text. That's weaker
-    (the widget may cover the transcript, same as on ChatGPT) — swap this out
-    for a real API check if/when one is found.
-    """
-    elapsed = 0
-    while elapsed < timeout_seconds:
-        found = page.evaluate(
-            "(marker) => document.body.innerText.includes(marker)",
-            FOLLOW_UP_MARKER,
-        )
-        if found:
-            return True
-        time.sleep(poll_interval)
-        elapsed += poll_interval
-    return False
-
-
-CLAUDE = HostConfig(
-    name="claude",
-    url="https://claude.ai/new",
-    # Verified live: Claude renders MCP Apps views in a sandboxed iframe on
-    # <hash>.claudemcpcontent.com/mcp_apps.
-    widget_iframe_selector='iframe[src*="claudemcpcontent"]',
-    send_prompt=send_prompt_claude,
-    hide_sidebar=hide_sidebar_claude,
-    dismiss_modal=dismiss_modal_claude,
-    check_follow_up=check_follow_up_claude,
-    accept_native_dialog=accept_native_dialog_claude,
+from chatgpt import CHATGPT
+from claude import CLAUDE
+from utils import (
+    MAX_LOOP_TURNS,
+    STEPPER_DEADLINE_S,
+    HostConfig,
+    ResultRow,
+    StallError,
+    _PENDING,
+    close_host_modal,
+    dismiss_host_overlay,
+    drive_widget,
+    install_state_listener,
+    read_state,
+    rows_complete,
+    take_screenshot,
+    wait_for_widget,
+    widget_present,
 )
 
 HOSTS: dict[str, HostConfig] = {CHATGPT.name: CHATGPT, CLAUDE.name: CLAUDE}
 
 
-# ── Widget interaction (host-agnostic) ───────────────────────────────
-
-
-def wait_for_widget(page: Page, selector: str, timeout_seconds: int = 90, poll_interval: int = 3) -> None:
-    """Poll until the widget iframe is in the conversation."""
-    elapsed = 0
-    while elapsed < timeout_seconds:
-        has_widget = page.evaluate("(sel) => Boolean(document.querySelector(sel))", selector)
-        if has_widget:
-            return
-        time.sleep(poll_interval)
-        elapsed += poll_interval
-    raise TimeoutError(f"widget iframe did not appear within {timeout_seconds}s")
-
-
-def center_widget(page: Page, selector: str) -> None:
-    """Scroll the top of the widget iframe into view.
-
-    The stepper card sits at the top of the app (results table below); with a
-    widget taller than the viewport, centering the iframe pushes the stepper
-    above the fold and the driver can no longer see the buttons.
-    """
-    page.evaluate(
-        """(sel) => {
-            const iframe = document.querySelector(sel);
-            if (iframe) iframe.scrollIntoView({ block: 'start', behavior: 'instant' });
-        }""",
-        selector,
-    )
-    time.sleep(1)
-
-
-def take_screenshot(page: Page, selector: str) -> bytes | None:
-    center_widget(page, selector)
-    # animations="disabled" finishes/cancels CSS animations first: Claude runs
-    # a perpetual one that otherwise hangs the default screenshot forever.
-    try:
-        return page.screenshot(animations="disabled", timeout=15_000)
-    except Exception as exc:
-        print(f"[conformance] screenshot failed: {exc}", flush=True)
-        return None
-
-
-# Button labels in the widget, per drive action (src/views/conformance.tsx).
-# The "run" action's label varies per test (Run/Close), so callers pass it.
-ACTION_LABELS = {
-    "yes": "Yes, it worked",
-    "no": "No effect",
-    "skip": "Skip",
-}
-
-
-# A real click's actionability checks (visible/stable/scroll) run over the wire,
-# so the timeout must cover round-trip latency: near-instant locally, but seconds
-# against a remote Notte cloud browser over CDP.
-CLICK_TIMEOUT_MS = 12_000
-
-
-def real_click(page: Page, label: str) -> bool:
-    """Click a widget button with a REAL, trusted mouse event.
-
-    The postMessage drive protocol carries no user activation, and ChatGPT
-    gates user-facing effects on a genuine gesture: a follow-up sent (or a
-    tab opened) from a synthetic click is silently dropped. Playwright can
-    click across the host's cross-origin iframes (something the Notte action
-    API could not), so prefer that and fall back to postMessage when the
-    button isn't reachable.
-
-    The widget's cross-origin frames churn while the host (re)mounts them,
-    especially over remote CDP, so a "Frame was detached" mid-scan is expected;
-    tolerate it and rescan a couple of times before giving up.
-    """
-    for attempt in range(3):
-        for frame in list(page.frames):
-            if frame is page.main_frame:
-                continue
-            try:
-                button = frame.get_by_role("button", name=label, exact=True)
-                if button.count():
-                    button.first.click(timeout=CLICK_TIMEOUT_MS)
-                    return True
-            except Exception as exc:
-                # First line only: Playwright's full click call-log is huge.
-                print(f"[conformance] real_click '{label}': {str(exc).splitlines()[0]}", flush=True)
-                continue
-        time.sleep(1)  # let a detached/remounting frame settle, then rescan
-    return False
-
-
-def drive_widget(page: Page, selector: str, action: str, label: str | None = None) -> None:
-    """Press a widget button: real click first, postMessage drive as fallback.
-
-    The widget lives behind cross-origin iframes (the host's sandboxed
-    origin, with a nested #root frame on some hosts); the app also listens
-    for {type: "conformance:drive", action} messages (actions: run, skip,
-    yes, no, close-modal, restore-inline). postMessage crosses origins, and
-    posting to the iframe window plus each of its child frames covers both
-    nesting layouts.
-    """
-    label = label or ACTION_LABELS.get(action)
-    if label and real_click(page, label):
-        print(f"[conformance] drive '{action}': clicked '{label}'", flush=True)
-        time.sleep(2)
-        return
-    result = page.evaluate(
-        """([sel, act]) => {
-            const iframe = document.querySelector(sel);
-            if (!iframe) return 'no-iframe';
-            const message = { type: 'conformance:drive', action: act };
-            const outer = iframe.contentWindow;
-            outer.postMessage(message, '*');
-            for (let i = 0; i < outer.frames.length; i++) {
-                outer.frames[i].postMessage(message, '*');
-            }
-            return 'sent';
-        }""",
-        [selector, action],
-    )
-    print(f"[conformance] drive '{action}': {result}", flush=True)
-    time.sleep(2)
-
-
-def install_state_listener(page: Page) -> None:
-    """Capture the app's state broadcasts on the host page.
-
-    The widget posts {type: "conformance:state", state} to window.top on
-    every change plus a 1.5s heartbeat; stash the latest payload (a real
-    object now, no JSON-string round trip needed) in a window variable that
-    read_state polls. Zero LLM involved.
-    """
-    page.evaluate(
-        """() => {
-            if (window.__confListenerInstalled) return 'already';
-            window.__confListenerInstalled = true;
-            window.addEventListener('message', (event) => {
-                if (event.data && event.data.type === 'conformance:state') {
-                    window.__confState = event.data.state;
-                }
-            });
-            return 'installed';
-        }"""
-    )
-
-
-def read_state(page: Page, timeout_seconds: int = 20, poll_interval: int = 2) -> StepperState:
-    """Read the latest state broadcast (deterministic replacement for scraping)."""
-    elapsed = 0
-    while elapsed < timeout_seconds:
-        data = page.evaluate("() => window.__confState || null")
-        if data:
-            return StepperState(**data)
-        time.sleep(poll_interval)
-        elapsed += poll_interval
-    raise TimeoutError(f"the widget broadcast no state within {timeout_seconds}s")
-
-
-def widget_present(page: Page, selector: str) -> bool:
-    return page.evaluate("(sel) => Boolean(document.querySelector(sel))", selector)
-
-
-# ── Outside verification of side effects ─────────────────────────────
-
-
-# A host-owned modal overlay, across both hosts: ChatGPT's dialog, and
-# Claude's base-ui / cds portal backdrops (which carry no role="dialog" but
-# blur the page and intercept every pointer event until dismissed).
-HOST_MODAL_SELECTOR = (
-    '[role="dialog"], [aria-modal="true"], '
-    "[data-cds-portal] [data-state=\"open\"], "
-    "[data-base-ui-portal] [data-open]"
-)
-
-
-def host_modal_present(page: Page) -> bool:
-    return page.evaluate("(sel) => Boolean(document.querySelector(sel))", HOST_MODAL_SELECTOR)
-
-
-def dismiss_host_overlay(page: Page, timeout_seconds: int = 12) -> None:
-    """Clear a host-owned modal/overlay and WAIT until it's actually gone.
-
-    The modal is host-owned on both hosts and the app's closeModal is a no-op
-    there. Its backdrop intercepts every pointer event, so if we don't wait
-    for it to clear the next test (useOpenExternal) fires into a live backdrop
-    that swallows its clicks and its new tab — on Claude this poisoned the
-    whole rest of the run. Escape is the host's own dismiss path, but the
-    close is async, so Escape then poll the top-page DOM, re-pressing until
-    the overlay is gone.
-    """
-    elapsed = 0
-    while elapsed < timeout_seconds:
-        try:
-            page.keyboard.press("Escape")
-        except Exception as exc:
-            print(f"[conformance] escape failed: {exc}", flush=True)
-            return
-        time.sleep(2)
-        elapsed += 2
-        if not host_modal_present(page):
-            return
-    print("[conformance] overlay still present after dismiss attempts", flush=True)
-
-
-def close_host_modal(page: Page, selector: str) -> None:
-    """Close whatever modal the requestModal test opened.
-
-    Escape covers host-chromed modals (ChatGPT), the drive command covers the
-    in-view polyfill Close on other hosts. Both are no-ops when nothing is
-    open.
-    """
-    try:
-        page.keyboard.press("Escape")
-        time.sleep(1)
-    except Exception:
-        pass
-    drive_widget(page, selector, "close-modal")
+# ── Answering confirmations ──────────────────────────────────────────
 
 
 def answer_confirm(
@@ -723,8 +228,8 @@ def drive_stepper(
             drive_widget(page, selector, "run", label=state.action_button.strip())
             if host.accept_native_dialog:
                 # Some hooks' effects are gated behind a host permission dialog
-                # (Claude: Open link / Download) that appears right after Run
-                # and blocks every later click until answered.
+                # (Open link / Download) that appears right after Run and blocks
+                # every later click until answered.
                 result = host.accept_native_dialog(page, hook)
                 if result is not None:
                     verified_effects[hook] = result
@@ -781,10 +286,9 @@ def run_conformance(
                 flush=True,
             )
     except Exception as exc:
-        # Swallow rather than propagate, same as the Notte version: a partial
-        # run (whatever rows got broadcast before things went wrong) is more
-        # useful than nothing, and the caller still gets a real result to
-        # write out and compare against the baseline.
+        # Swallow rather than propagate: a partial run (whatever rows got
+        # broadcast before things went wrong) is more useful than nothing, and
+        # the caller still gets a real result to write out and compare.
         print(f"[conformance] stepper run failed: {exc}", flush=True)
         if screenshot_bytes is None:
             try:
@@ -823,9 +327,7 @@ def local_browser(profile_dir: Path) -> Iterator[BrowserContext]:
             viewport={"width": 1440, "height": 1400},
             # navigator.webdriver=true trips the hosts' bot detection
             # (Cloudflare on auth.openai.com); this is our own test account.
-            # Popup blocking must be off: the openExternal test's new tab is
-            # opened without a user gesture (the click is driven via
-            # postMessage), which the blocker would silently eat.
+            # Popup blocking must be off so the openExternal tab isn't eaten.
             args=[
                 "--disable-blink-features=AutomationControlled",
                 "--disable-popup-blocking",
@@ -909,19 +411,6 @@ UNVERIFIABLE_BY_MODE: dict[str, frozenset[str]] = {}
 # ── CLI ───────────────────────────────────────────────────────────────
 
 
-def load_dotenv(path: Path) -> None:
-    """Populate os.environ from a .env file (KEY=VALUE lines) without clobbering
-    existing vars. Enough for the notte:* vars; CI passes real env vars instead."""
-    if not path.exists():
-        return
-    for line in path.read_text().splitlines():
-        line = line.strip()
-        if not line or line.startswith("#") or "=" not in line:
-            continue
-        key, _, value = line.partition("=")
-        os.environ.setdefault(key.strip(), value.strip())
-
-
 def parse_args(argv: list[str] | None = None) -> argparse.Namespace:
     script_dir = Path(__file__).resolve().parent
     parser = argparse.ArgumentParser(
@@ -965,8 +454,8 @@ def parse_args(argv: list[str] | None = None) -> argparse.Namespace:
 
 def main(argv: list[str] | None = None) -> int:
     # Load the example's .env (NOTTE_API_KEY / NOTTE_PROFILE_ID) before args are
-    # parsed, so the --profile-id default and NotteClient both see it. CI sets
-    # these as real env vars, so setdefault leaves them untouched there.
+    # parsed, so the --profile-id default and NotteClient both see it. override
+    # is False by default, so CI's real env vars win over the file.
     load_dotenv(Path(__file__).resolve().parent.parent / ".env")
 
     args = parse_args(argv)
